rename: strategies폴더에서 (dashboard)/_ui 하위로 이동 (#47)
