+++ app/(dashboard)/_ui/strategies-item/strategies-icon.tsx
@@ -1,6 +1,7 @@
 interface Props {
   icon: string[] | React.ElementType
 }
+
 const StrategiesIcon = ({ icon }: Props) => {
   return <div>[종목 & 매매 Icon]</div>
 }
