+++ app/(dashboard)/_ui/strategies-item/styles.module.scss
@@ -18,7 +18,7 @@
   }
   .profit {
     & span {
-      @include typo-c2;
+      @include typo-c1;
     }
     p {
       @include typo-b2;
@@ -48,7 +48,7 @@
     align-items: center;
     height: 20px;
     p {
-      @include typo-c2;
+      @include typo-c1;
       padding-top: 10px;
       &:first-child {
         margin-right: 6px;
