+++ app/(dashboard)/_ui/strategies-item/subscribe.tsx
